+++ opencosmo/collection/collection.py
@@ -7,8 +7,8 @@ from astropy.cosmology import Cosmology  # type: ignore
 
 import opencosmo as oc
 from opencosmo.collection.protocols import Collection
+from opencosmo.dataset.col import Mask
 from opencosmo.dataset.handler import DatasetHandler
-from opencosmo.dataset.mask import Mask
 from opencosmo.dataset.state import DatasetState
 from opencosmo.header import OpenCosmoHeader, read_header
 from opencosmo.index import ChunkedIndex
