+++ opencosmo/dataset/__init__.py
@@ -1,4 +1,4 @@
+from .col import col
 from .dataset import Dataset
-from .mask import col
 
 __all__ = ["col", "Dataset"]
